borrow in new structure
