+++ features/ajna/borrow/controls/AjnaBorrowOverviewController.tsx
@@ -6,9 +6,9 @@ import { ContentCardLiquidationPrice } from 'features/ajna/borrow/overview/Conte
 import { ContentCardLoanToValue } from 'features/ajna/borrow/overview/ContentCardLoanToValue'
 import { ContentCardPositionDebt } from 'features/ajna/borrow/overview/ContentCardPositionDebt'
 import { ContentFooterItemsBorrow } from 'features/ajna/borrow/overview/ContentFooterItemsBorrow'
-import { useAjnaGeneralContext } from 'features/ajna/contexts/AjnaGeneralContext'
-import { useAjnaProductContext } from 'features/ajna/contexts/AjnaProductContext'
-import { AjnaTokensBanner } from 'features/ajna/controls/AjnaTokensBannerController'
+import { useAjnaGeneralContext } from 'features/ajna/common/contexts/AjnaGeneralContext'
+import { useAjnaProductContext } from 'features/ajna/common/contexts/AjnaProductContext'
+import { AjnaTokensBanner } from 'features/ajna/common/controls/AjnaTokensBannerController'
 import { useTranslation } from 'next-i18next'
 import React from 'react'
 import { Grid } from 'theme-ui'
