+++ features/ajna/borrow/sidebars/AjnaBorrowFormContentSummary.tsx
@@ -1,7 +1,7 @@
 import { SidebarResetButton } from 'components/vault/sidebar/SidebarResetButton'
 import { AjnaBorrowFormOrder } from 'features/ajna/borrow/sidebars/AjnaBorrowFormOrder'
-import { AjnaValidationMessages } from 'features/ajna/components/AjnaValidationMessages'
-import { useAjnaProductContext } from 'features/ajna/contexts/AjnaProductContext'
+import { AjnaValidationMessages } from 'features/ajna/common/components/AjnaValidationMessages'
+import { useAjnaProductContext } from 'features/ajna/common/contexts/AjnaProductContext'
 import React from 'react'
 
 export function AjnaBorrowFormContentSummary() {
