+++ features/ajna/common/contexts/AjnaProductContext.tsx
@@ -3,9 +3,9 @@ import { useAppContext } from 'components/AppContextProvider'
 import { useGasEstimationContext } from 'components/GasEstimationContextProvider'
 import { ValidationMessagesInput } from 'components/ValidationMessages'
 import { useAjnaBorrowFormReducto } from 'features/ajna/borrow/state/ajnaBorrowFormReducto'
+import { useAjnaGeneralContext } from 'features/ajna/common/contexts/AjnaGeneralContext'
 import { AjnaProduct } from 'features/ajna/common/types'
 import { getAjnaValidation } from 'features/ajna/common/validation'
-import { useAjnaGeneralContext } from 'features/ajna/contexts/AjnaGeneralContext'
 import { AjnaEarnPosition } from 'features/ajna/earn/fakePosition'
 import { useAjnaEarnFormReducto } from 'features/ajna/earn/state/ajnaEarnFormReducto'
 import { useObservable } from 'helpers/observableHook'
