+++ features/ajna/common/controls/AjnaProductController.tsx
@@ -4,11 +4,11 @@ import { PositionLoadingState } from 'components/vault/PositionLoadingState'
 import { AjnaBorrowPositionController } from 'features/ajna/borrow/controls/AjnaBorrowPositionController'
 import { useAjnaBorrowFormReducto } from 'features/ajna/borrow/state/ajnaBorrowFormReducto'
 import { steps } from 'features/ajna/common/consts'
-import { getAjnaHeadlineProps } from 'features/ajna/common/helpers'
+import { AjnaGeneralContextProvider } from 'features/ajna/common/contexts/AjnaGeneralContext'
+import { AjnaProductContextProvider } from 'features/ajna/common/contexts/AjnaProductContext'
+import { getAjnaHeadlineProps } from 'features/ajna/common/helpers/getAjnaHeadlineProps'
 import { AjnaWrapper } from 'features/ajna/common/layout'
 import { AjnaFlow, AjnaProduct } from 'features/ajna/common/types'
-import { AjnaGeneralContextProvider } from 'features/ajna/contexts/AjnaGeneralContext'
-import { AjnaProductContextProvider } from 'features/ajna/contexts/AjnaProductContext'
 import { ajnaPositionToAjnaEarnPosition } from 'features/ajna/earn/fakePosition'
 import { useAjnaEarnFormReducto } from 'features/ajna/earn/state/ajnaEarnFormReducto'
 import { WithTermsOfService } from 'features/termsOfService/TermsOfService'
