+++ features/ajna/common/helpers/getAjnaSidebarButtonsStatus.tsx
@@ -1,4 +1,4 @@
-import { AjnaFlow, AjnaSidebarEditingStep, AjnaSidebarStep } from 'features/ajna/common/types'
+import { AjnaSidebarEditingStep, AjnaSidebarStep } from 'features/ajna/common/types'
 
 export function getAjnaSidebarButtonsStatus({
   currentStep,
@@ -33,14 +33,6 @@ export function getAjnaSidebarButtonsStatus({
       isTxInProgress ||
       isTxWaitingForApproval)
 
-  console.log('---')
-  console.log(`walletAddress: ${walletAddress}`)
-  console.log(`isFormValid: ${isFormValid}`)
-  console.log(`isAllowanceLoading: ${isAllowanceLoading}`)
-  console.log(`isSimulationLoading: ${isSimulationLoading}`)
-  console.log(`isTxInProgress: ${isTxInProgress}`)
-  console.log(`isTxWaitingForApproval: ${isTxWaitingForApproval}`)
-
   const isPrimaryButtonLoading =
     !!walletAddress &&
     (isAllowanceLoading || isSimulationLoading || isTxInProgress || isTxWaitingForApproval)
@@ -53,32 +45,5 @@ export function getAjnaSidebarButtonsStatus({
     isPrimaryButtonHidden,
     isPrimaryButtonLoading,
     isTextButtonHidden,
-  }
-}
-
-export function getAjnaSidebarPrimaryButtonActions({
-  currentStep,
-  defaultAction,
-  editingStep,
-  flow,
-  resolvedId,
-  isTxSuccess,
-  walletAddress,
-}: {
-  currentStep: string
-  defaultAction: () => void
-  editingStep: string
-  flow: AjnaFlow
-  resolvedId?: string
-  isTxSuccess: boolean
-  walletAddress?: string
-}) {
-  switch (true) {
-    case !walletAddress && currentStep === editingStep:
-      return { url: '/connect' }
-    case isTxSuccess && flow === 'open':
-      return { url: `/ajna/position/${resolvedId}` }
-    default:
-      return { action: defaultAction }
   }
 }
