+++ features/ajna/common/helpers/getAjnaWithArrowColorScheme.ts
@@ -1,0 +1,9 @@
+import { SxStyleProp } from 'theme-ui'
+
+export function getAjnaWithArrowColorScheme(): SxStyleProp {
+  return {
+    color: 'interactive100',
+    transition: 'color 200ms',
+    '&:hover': { color: 'interactive50' },
+  }
+}
