+++ features/ajna/common/views/AjnaFormView.tsx
@@ -2,15 +2,13 @@ import { FlowSidebar } from 'components/FlowSidebar'
 import { SidebarSection, SidebarSectionProps } from 'components/sidebar/SidebarSection'
 import { SidebarSectionHeaderDropdown } from 'components/sidebar/SidebarSectionHeader'
 import { ethers } from 'ethers'
-import {
-  getAjnaSidebarButtonsStatus,
-  getAjnaSidebarPrimaryButtonActions,
-} from 'features/ajna/borrow/helpers'
-import { useAjnaTxHandler } from 'features/ajna/borrow/useAjnaTxHandler'
-import { getPrimaryButtonLabelKey } from 'features/ajna/common/helpers'
+import { useAjnaGeneralContext } from 'features/ajna/common/contexts/AjnaGeneralContext'
+import { useAjnaProductContext } from 'features/ajna/common/contexts/AjnaProductContext'
+import { getAjnaSidebarButtonsStatus } from 'features/ajna/common/helpers/getAjnaSidebarButtonsStatus'
+import { getAjnaSidebarPrimaryButtonActions } from 'features/ajna/common/helpers/getAjnaSidebarPrimaryButtonActions'
 import { getFlowStateConfig } from 'features/ajna/common/helpers/getFlowStateConfig'
-import { useAjnaGeneralContext } from 'features/ajna/contexts/AjnaGeneralContext'
-import { useAjnaProductContext } from 'features/ajna/contexts/AjnaProductContext'
+import { getPrimaryButtonLabelKey } from 'features/ajna/common/helpers/getPrimaryButtonLabelKey'
+import { useAjnaTxHandler } from 'features/ajna/common/hooks/useAjnaTxHandler'
 import { useAccount } from 'helpers/useAccount'
 import { useFlowState } from 'helpers/useFlowState'
 import { useTranslation } from 'next-i18next'
