+++ features/ajna/controls/AjnaSelectorController.tsx
@@ -5,7 +5,7 @@ import { HeaderSelector, HeaderSelectorOption } from 'components/HeaderSelector'
 import { AppLink } from 'components/Links'
 import { ajnaComingSoonPools, DEFAULT_SELECTED_TOKEN } from 'features/ajna/common/consts'
 import { ajnaPoolDummyData } from 'features/ajna/common/content'
-import { filterPoolData } from 'features/ajna/common/helpers'
+import { filterPoolData } from 'features/ajna/common/helpers/filterPoolData'
 import { AjnaWrapper } from 'features/ajna/common/layout'
 import { AjnaProduct } from 'features/ajna/common/types'
 import { DiscoverResponsiveTable } from 'features/discover/common/DiscoverResponsiveTable'
